Way faster tests, fixed Picture tests at the same time
