+++ app/src/androidTest/java/com/github/factotum_sdp/factotum/ui/picture/utils/PictureFragmentTestHelper.kt
@@ -1,5 +1,6 @@
 package com.github.factotum_sdp.factotum.ui.picture
 
+import android.util.Log
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.swipeLeft
@@ -9,15 +10,20 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.uiautomator.UiDevice
 import com.github.factotum_sdp.factotum.R
 import com.github.factotum_sdp.factotum.placeholder.DestinationRecords
+import com.github.factotum_sdp.factotum.placeholder.UsersPlaceHolder
 import com.github.factotum_sdp.factotum.ui.login.LoginFragmentTest
 import com.github.factotum_sdp.factotum.utils.PreferencesSetting
+import com.google.firebase.auth.FirebaseAuthUserCollisionException
 import com.google.firebase.storage.StorageReference
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.tasks.await
 import java.io.File
 
-const val TIME_WAIT_SHUTTER = 5000L
-const val TIME_WAIT_DONE_OR_CANCEL = 3000L
-const val TIME_WAIT_UPLOAD = 500L
+const val TIME_WAIT_SHUTTER = 1000L
+const val TIME_WAIT_DONE_OR_CANCEL = 1000L
+const val TIME_WAIT_UPLOAD_PHOTO = 1000L
+const val CLIENT_ID = "X17"
 
 
 // HELPER METHODS
@@ -63,12 +69,6 @@ fun triggerCancel(device: UiDevice) {
 fun goToPictureFragment() {
     LoginFragmentTest.fillUserEntryAndGoToRBFragment("courier@gmail.com", "123456")
 
-    // Open the drawer
-    onView(withId(R.id.drawer_layout))
-        .perform(open())
-    onView(withId(R.id.roadBookFragment))
-        .perform(click())
-
     PreferencesSetting.enableTouchClick()
 
     // Click on one of the roadbook
@@ -79,4 +79,17 @@ fun goToPictureFragment() {
     onView(withId(R.id.viewPager)).perform(swipeLeft())
     onView(withId(R.id.viewPager)).perform(swipeLeft())
     onView(withId(R.id.viewPager)).perform(swipeLeft())
+}
+
+fun addUserToDatabase(user : UsersPlaceHolder.User) {
+    // DO NOT REMOVE THIS PART OR PUT IT IN A @BeforeClass
+    runBlocking {
+        try {
+            UsersPlaceHolder.addAuthUser(user)
+        } catch (e : FirebaseAuthUserCollisionException) {
+            e.message?.let { Log.e("DisplayFragmentTest", it) }
+        }
+
+        UsersPlaceHolder.addUserToDb(user)
+    }
 }
